+++ docs/en/changes/changes.md
@@ -50,6 +50,7 @@
 * Support MySQL monitoring.
 * Fix query services by serviceId error when Elasticsearch storage `SW_STORAGE_ES_QUERY_MAX_SIZE` > 10000.
 * Support sending alarm messages to Discord.
+* Fix query history process data failure.
 
 #### UI